+++ checkpoint/orbax/checkpoint/test_utils.py
@@ -45,6 +45,7 @@ from orbax.checkpoint._src.multihost import multihost
 from orbax.checkpoint._src.multihost import multislice
 from orbax.checkpoint._src.path import atomicity
 from orbax.checkpoint._src.path import step as step_lib
+from orbax.checkpoint._src.serialization import replica_slices
 from orbax.checkpoint._src.serialization import serialization
 from orbax.checkpoint._src.serialization import tensorstore_utils as ts_utils
 from orbax.checkpoint._src.serialization import type_handlers
@@ -736,15 +737,9 @@ def assert_every_n_is_x_apart(testclass, values, n, x):
 
 def get_expected_chunk_shape(arr: jax.Array) -> tuple[int, ...]:
   """Expected chunk shape for an array, accounting for replica-parallel."""
-  # TODO(cpgaffney): Enable once replica-parallel is enabled.
-  local_shape = None
-  # get_replica_counts = replica_slices._create_replica_counts_builder(arr)
-  # replica_count = get_replica_counts(arr.addressable_shards[0])
-  # _, local_shape = (
-  #     replica_slices.calculate_replica_parallel_axis_and_local_shape(
-  #         arr, replica_count
-  #     )
-  # )
+  _, local_shape = (
+      replica_slices.calculate_replica_parallel_axis_and_local_shape(arr)
+  )
   if local_shape is None:
     local_shape = arr.sharding.shard_shape(arr.shape)
   return local_shape